# BackEnd/App/docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:53181","description":"Generated server url"}],"paths":{"/auth/reset-password":{"post":{"tags":["password-reset-controller"],"operationId":"cambiarPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/auth/forgot-password":{"post":{"tags":["password-reset-controller"],"operationId":"solicitarRecuperacion","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/registro/profesional":{"post":{"tags":["registro-controller"],"operationId":"registrarProfesional","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NuevoProfesional"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Profesionale"}}}}}}},"/api/v1/registro/cliente":{"post":{"tags":["registro-controller"],"operationId":"registrarCliente","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NuevoCliente"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Cliente"}}}}}}},"/api/v1/oficios/all":{"get":{"tags":["oficio-controller"],"operationId":"getAllOficios","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Oficio"}}}}}}}},"/api/v1/departamentos/all":{"get":{"tags":["departamento-controller"],"operationId":"getAllOficios_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Departamento"}}}}}}}}},"components":{"schemas":{"ResetPasswordRequest":{"type":"object","properties":{"email":{"type":"string"},"codigo":{"type":"string"},"nuevaPassword":{"type":"string"}}},"RegisterRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"},"name":{"type":"string"},"lastname":{"type":"string"}}},"AuthResponse":{"type":"object","properties":{"token":{"type":"string"}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"ForgotPasswordRequest":{"type":"object","properties":{"email":{"type":"string"}}},"LocalTime":{"type":"object","properties":{"hour":{"type":"integer","format":"int32"},"minute":{"type":"integer","format":"int32"},"second":{"type":"integer","format":"int32"},"nano":{"type":"integer","format":"int32"}}},"NuevoProfesional":{"type":"object","properties":{"idUsuario":{"type":"integer","format":"int32"},"idTipoDoc":{"type":"integer","format":"int32"},"idoficio":{"type":"integer","format":"int32"},"documento":{"type":"string"},"telefono":{"type":"string"},"fechaNacimiento":{"$ref":"#/components/schemas/LocalTime"},"fechaDesde":{"$ref":"#/components/schemas/LocalTime"},"fechaHasta":{"$ref":"#/components/schemas/LocalTime"},"idDireccion":{"type":"integer","format":"int32"}}},"Barrio":{"required":["barrio","idciudad"],"type":"object","properties":{"id":{"type":"integer","format":"int32"},"barrio":{"maxLength":100,"minLength":0,"type":"string"},"idciudad":{"$ref":"#/components/schemas/Ciudade"}}},"Ciudade":{"required":["ciudad","iddepartamento"],"type":"object","properties":{"id":{"type":"integer","format":"int32"},"ciudad":{"maxLength":100,"minLength":0,"type":"string"},"iddepartamento":{"$ref":"#/components/schemas/Departamento"}}},"Departamento":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"departamento":{"maxLength":255,"minLength":0,"type":"string"}}},"Direccione":{"required":["calle","idbarrio","idusuario","numero"],"type":"object","properties":{"id":{"type":"integer","format":"int32"},"idusuario":{"$ref":"#/components/schemas/Usuario"},"idbarrio":{"$ref":"#/components/schemas/Barrio"},"calle":{"maxLength":100,"minLength":0,"type":"string"},"numero":{"maxLength":10,"minLength":0,"type":"string"},"piso":{"maxLength":10,"minLength":0,"type":"string"},"depto":{"maxLength":10,"minLength":0,"type":"string"},"observaciones":{"maxLength":200,"minLength":0,"type":"string"}}},"GrantedAuthority":{"type":"object","properties":{"authority":{"type":"string"}}},"Oficio":{"required":["oficio"],"type":"object","properties":{"id":{"type":"integer","format":"int32"},"oficio":{"maxLength":100,"minLength":0,"type":"string"}}},"Profesionale":{"required":["fechadesde","iddireccion","idoficio","idtipodoc","idusuario","nacimiento"],"type":"object","properties":{"id":{"type":"integer","format":"int32"},"idtipodoc":{"$ref":"#/components/schemas/TiposDocumento"},"documento":{"maxLength":20,"minLength":0,"type":"string"},"telefono":{"maxLength":20,"minLength":0,"type":"string"},"idusuario":{"$ref":"#/components/schemas/Usuario"},"idoficio":{"$ref":"#/components/schemas/Oficio"},"nacimiento":{"$ref":"#/components/schemas/LocalTime"},"fechadesde":{"$ref":"#/components/schemas/LocalTime"},"fechahasta":{"$ref":"#/components/schemas/LocalTime"},"iddireccion":{"$ref":"#/components/schemas/Direccione"}}},"Role":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"descripcion":{"type":"string"},"usuarios":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/Usuario"}}}},"TiposDocumento":{"required":["tipo"],"type":"object","properties":{"id":{"type":"integer","format":"int32"},"tipo":{"maxLength":255,"minLength":0,"type":"string"}}},"Usuario":{"required":["active","lastname","mail","name","password"],"type":"object","properties":{"id":{"type":"integer","format":"int32"},"password":{"maxLength":255,"minLength":0,"type":"string"},"name":{"maxLength":255,"minLength":0,"type":"string"},"lastname":{"maxLength":255,"minLength":0,"type":"string"},"mail":{"maxLength":150,"minLength":0,"type":"string"},"active":{"type":"boolean"},"roles":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/Role"}},"enabled":{"type":"boolean"},"accountNonLocked":{"type":"boolean"},"credentialsNonExpired":{"type":"boolean"},"accountNonExpired":{"type":"boolean"},"username":{"type":"string"},"authorities":{"type":"array","items":{"$ref":"#/components/schemas/GrantedAuthority"}}}},"NuevoCliente":{"type":"object","properties":{"idUsuario":{"type":"integer","format":"int32"},"idTipoDoc":{"type":"integer","format":"int32"},"documento":{"type":"string"},"telefono":{"type":"string"},"fechaNacimiento":{"$ref":"#/components/schemas/LocalTime"},"idDireccion":{"type":"integer","format":"int32"}}},"Cliente":{"required":["iddireccion","idtipodoc","idusuario","nacimiento"],"type":"object","properties":{"id":{"type":"integer","format":"int32"},"idtipodoc":{"$ref":"#/components/schemas/TiposDocumento"},"documento":{"maxLength":20,"minLength":0,"type":"string"},"telefono":{"maxLength":20,"minLength":0,"type":"string"},"nacimiento":{"$ref":"#/components/schemas/LocalTime"},"idusuario":{"$ref":"#/components/schemas/Usuario"},"iddireccion":{"$ref":"#/components/schemas/Direccione"}}}}}}